Improve configuration, styles, and documentation
Changes:
  - Configuration
      So the footer can be even more custom I've made each element
inside it option, and changeable, also I've removed the Git link there,
since we have one in the nav. By that I've also added a new option to
the theme called "author", this one is mostly used for the copyright.
I've also formatted the nav (not fully sure of it).
  - Overall better styles
  - Changed some words in the documentation.